# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /strategies:
    get:
      description: Returns paginated list of strategies for the authenticated user with optional filters
      operationId: StrategyController_listStrategies
      parameters:
        - name: trading_account_id
          required: false
          in: query
          description: Filter by trading account ID
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: mode
          required: false
          in: query
          description: Filter by trading mode
          schema:
            type: string
            enum:
              - REAL
              - PAPER
        - name: is_live
          required: false
          in: query
          description: Filter by live status
          schema:
            example: true
            type: boolean
        - name: include_archived
          required: false
          in: query
          description: Include archived strategies in results
          schema:
            default: false
            example: false
            type: boolean
        - name: archived_only
          required: false
          in: query
          description: Show only archived strategies
          schema:
            default: false
            example: false
            type: boolean
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination (strategy ID)
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Strategies retrieved successfully
        '401':
          description: Unauthorized
      security:
        - JWT-auth: []
      summary: List user strategies
      tags: &ref_5
        - Strategies
    post:
      description: Creates a new trading strategy and an initial portfolio. The portfolio type matches the strategy mode (PAPER by default).
      operationId: StrategyController_createStrategy
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStrategyDto'
      responses:
        '201':
          description: Strategy created successfully
        '401':
          description: Unauthorized
        '404':
          description: Trading account not found
        '409':
          description: Strategy with this name already exists for this account
      security:
        - JWT-auth: []
      summary: Create new strategy
      tags: *ref_5
  /strategies/public:
    get:
      description: Returns paginated list of public strategies from all users. User data is anonymized.
      operationId: StrategyController_listPublicStrategies
      parameters:
        - name: search
          required: false
          in: query
          description: Search by strategy name or description
          schema:
            example: double bottom
            type: string
        - name: sort
          required: false
          in: query
          description: Sort order
          schema:
            default: newest
            type: string
            enum:
              - popular
              - performance
              - newest
        - name: cursor
          required: false
          in: query
          description: Cursor for pagination (strategy ID)
          schema:
            example: clxy1234567890abcdefghijk
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            example: 20
            type: number
      responses:
        '200':
          description: Public strategies retrieved successfully
      summary: Browse public strategies marketplace
      tags: *ref_5
  /strategies/{id}:
    get:
      description: Returns detailed information about a strategy including portfolios and trading account. User must own the strategy OR strategy must be public.
      operationId: StrategyController_getStrategyById
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy retrieved successfully
        '401':
          description: Unauthorized
        '403':
          description: You do not have access to this strategy
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Get strategy details
      tags: *ref_5
    patch:
      description: Updates strategy configuration. Cannot change mode while is_live=true. Cannot switch to REAL mode without a REAL portfolio.
      operationId: StrategyController_updateStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStrategyDto'
      responses:
        '200':
          description: Strategy updated successfully
        '400':
          description: Invalid update - check validation rules
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
        '409':
          description: Strategy with this name already exists for this account
      security:
        - JWT-auth: []
      summary: Update strategy settings
      tags: *ref_5
    delete:
      description: Deletes a strategy. Cannot delete if is_live=true or if there are open positions. This action is permanent.
      operationId: StrategyController_deleteStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy deleted successfully
        '400':
          description: Cannot delete - strategy is live or has open positions
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Delete strategy
      tags: *ref_5
  /strategies/{id}/copy:
    post:
      description: Creates a copy of a public strategy linked to your trading account. The copied strategy starts in PAPER mode and is_live=false.
      operationId: StrategyController_copyStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Source strategy ID to copy
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CopyStrategyDto'
      responses:
        '201':
          description: Strategy copied successfully
        '401':
          description: Unauthorized
        '403':
          description: Source strategy is not public
        '404':
          description: Strategy or trading account not found
        '409':
          description: Strategy with this name already exists for this account
      security:
        - JWT-auth: []
      summary: Copy a public strategy
      tags: *ref_5
  /strategies/{id}/archive:
    post:
      description: Archives a strategy. Cannot archive if is_live=true. Archived strategies are hidden by default but can be filtered in list endpoints.
      operationId: StrategyController_archiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy archived successfully
        '400':
          description: Cannot archive - strategy is live or already archived
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Archive strategy
      tags: *ref_5
  /strategies/{id}/unarchive:
    post:
      description: Unarchives a strategy. Strategy must be archived.
      operationId: StrategyController_unarchiveStrategy
      parameters:
        - name: id
          required: true
          in: path
          description: Strategy ID
          schema:
            type: string
      responses:
        '200':
          description: Strategy unarchived successfully
        '400':
          description: Cannot unarchive - strategy is not archived
        '401':
          description: Unauthorized
        '404':
          description: Strategy not found
      security:
        - JWT-auth: []
      summary: Unarchive strategy
      tags: *ref_5
...
info:
  title: Crypto Watcher Bot API
  description: API for crypto trading signal bot with automated order execution
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    JWT-auth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      name: JWT
      description: Enter JWT token
      in: header
  schemas:
...
    CreateStrategyDto:
      type: object
      properties:
        name:
          type: string
          description: Strategy name
          example: My Double Bottom Strategy
        description:
          type: string
          description: Strategy description
          example: Conservative double bottom pattern strategy
        trading_account_id:
          type: string
          description: Trading account ID
          example: clxy1234567890abcdefghijk
        trade_size_type:
          type: string
          description: Trade size type
          enum:
            - PERCENTAGE
            - USD
          default: PERCENTAGE
        trade_size_amount:
          type: number
          description: Trade size amount
          example: 5
          default: 5
        trade_size_benchmark:
          type: string
          description: Trade size benchmark
          enum:
            - BALANCE
            - SL
            - TP
          default: SL
        min_risk_ratio:
          type: number
          description: Minimum risk ratio filter
          example: 1.5
        max_risk_ratio:
          type: number
          description: Maximum risk ratio filter
          example: 10
        allowed_signals:
          description: Allowed signal patterns
          example:
            - double_bottom
            - double_top
          items:
            type: array
          type: array
        allowed_symbols:
          description: Allowed trading symbols
          example:
            - BTCUSDT
            - ETHUSDT
          items:
            type: array
          type: array
        allowed_timeframes:
          description: Allowed timeframes
          example:
            - 1m
            - 15m
            - 1h
            - 4h
          items:
            type: array
          type: array
        mode:
          type: string
          description: Trading mode (REAL or PAPER)
          enum:
            - REAL
            - PAPER
          default: PAPER
        is_live:
          type: boolean
          description: Is strategy currently running
          default: false
        is_public:
          type: boolean
          description: Is strategy visible to other users
          default: false
        initial_capital:
          type: number
          description: Initial capital for portfolio
          example: 10000
      required:
        - name
        - trading_account_id
        - initial_capital
    CopyStrategyDto:
      type: object
      properties:
        trading_account_id:
          type: string
          description: Trading account ID to link the copied strategy to
          example: clxy1234567890abcdefghijk
        name:
          type: string
          description: Custom name for the copied strategy (defaults to "Copy of {original}")
          example: My Custom Strategy Name
      required:
        - trading_account_id
    UpdateStrategyDto:
      type: object
      properties:
        name:
          type: string
          description: Strategy name
          example: My Updated Strategy
        description:
          type: string
          description: Strategy description
          example: Updated strategy description
        is_live:
          type: boolean
          description: Is strategy currently running
        mode:
          type: string
          description: Trading mode (REAL or PAPER)
          enum:
            - REAL
            - PAPER
        is_public:
          type: boolean
          description: Is strategy visible to other users
        trade_size_type:
          type: string
          description: Trade size type
          enum:
            - PERCENTAGE
            - USD
        trade_size_amount:
          type: number
          description: Trade size amount
          example: 5
        trade_size_benchmark:
          type: string
          description: Trade size benchmark
          enum:
            - BALANCE
            - SL
            - TP
        min_risk_ratio:
          type: number
          description: Minimum risk ratio filter
          example: 1.5
        max_risk_ratio:
          type: number
          description: Maximum risk ratio filter
          example: 10
        allowed_signals:
          description: Allowed signal patterns
          example:
            - double_bottom
            - double_top
          items:
            type: array
          type: array
        allowed_symbols:
          description: Allowed trading symbols
          example:
            - BTCUSDT
            - ETHUSDT
          items:
            type: array
          type: array
        allowed_timeframes:
          description: Allowed timeframes
          example:
            - 1m
            - 15m
            - 1h
            - 4h
          items:
            type: array
          type: array
        is_active:
          type: boolean
          description: Is strategy active
...